refactor(inspect): move inspect module out of util. Rename script
