+++ falcon/inspect.py
