+++ setup.py
@@ -173,7 +173,7 @@ setup(
     entry_points={
         'console_scripts': [
             'falcon-bench = falcon.cmd.bench:main',
-            'falcon-print-routes = falcon.cmd.print_routes:main'
+            'falcon-inspect-app = falcon.cmd.inspect_app:main'
         ]
     }
 )
